+++ include/common/strings.h
@@ -6,5 +6,5 @@
 
 #include <common/stddef.h>
 
-void bzero(void *dest, int bytes);
+void bzero(void *dest, size_t bytes);
 void bcopy(const void *src, void *dest, size_t bytes);
